Revert "Vet UI rebased (#32)" (#33)
This reverts commit 60a8866f6eac796f9ee9d37458cf2e2b91b80494.
