--- lifu/__init__.py
@@ -1,3 +1,0 @@
-from .lifu_connector import LIFUConnector, MIN_SDK_VERSION, check_sdk_version
-
-__all__ = ["LIFUConnector", "MIN_SDK_VERSION", "check_sdk_version"]
