+++ pages/Console.qml
@@ -22,6 +22,7 @@ Rectangle {
     property string v12State: "Off" // Add property for 12V state
 
     function updateStates() {
+        console.log("Updating all states...")
         LIFUConnector.queryHvInfo()
         LIFUConnector.queryHvTemperature()
         LIFUConnector.queryPowerStatus() // Query power status
@@ -41,6 +42,7 @@ Rectangle {
         interval: 500   // Delay to ensure HV is stable before fetching info
         running: false
         onTriggered: {
+            console.log("Fetching Firmware Version and Device ID...")
             updateStates()
         }
     }
@@ -53,13 +55,13 @@ Rectangle {
             if (LIFUConnector.hvConnected) {
                 infoTimer.start()          // One-time info fetch
             } else {
+                console.log("HV Disconnected - Clearing Data...")
                 firmwareVersion = "N/A"
                 deviceId = "N/A"
                 temperature1 = 0.0
                 temperature2 = 0.0
                 rgbState = "Off" // Reset RGB state
-                hvState = "Off"
-                v12State = "Off"
+                voltageState = "Off" // Reset voltage state
                 pingResult.text = ""
                 echoResult.text = ""
                 toggleLedResult.text = ""
