+++ pages/Transmitter.qml
@@ -17,6 +17,7 @@ Rectangle {
     property bool loading: false
 
     function updateStates() {
+        console.log("Updating all states...")
         loading = true
         queryTimer.start()
     }
@@ -34,6 +35,7 @@ Rectangle {
         interval: 1500   // Delay to ensure TX is stable before fetching info
         running: false
         onTriggered: {
+            console.log("Fetching Firmware Version and Device ID...")
             updateStates()
         }
     }
