--- plot/__init__.py
@@ -1,9 +1,0 @@
-from .plot import (
-    generate_ultrasound_plot,
-    generate_ultrasound_plot_from_solution,
-)
-
-__all__ = [
-    "generate_ultrasound_plot",
-    "generate_ultrasound_plot_from_solution",
-]
