+++ scripts/generate_ultrasound_plot.py
