--- test_reports/__init__.py
@@ -1,11 +1,0 @@
-from .test_reports import (
-    read_test_report,
-    test_report_to_config,
-    check_config_against_device,
-)
-
-__all__ = [
-    "read_test_report",
-    "test_report_to_config",
-    "check_config_against_device",
-]
